Split serializer out into its own file.
